requirements.txt
flask
mcipc
psutil
distro
pyyaml
requests
markdown
Turbo-Flask